Added more network annotations
This extracts more logic into the setting of our network-related
annotations to replace the "legacy" ones.

When we start using these new annotations, they can be taken "as is"
without any additional logic by the control-plane or anyone else looking
at the pod. (and we can deprecate the legacy annotations)