Removed last use of TITUS_TASK_INSTANCE_ID
I hope I'm not missing any historical context about this variable.
It seems redundant in the face of TITUS_TASK_ID.

This PR removes the last trace of it, as it looks like it was only for
the imds.

It is not a documented var on titus-docs, customers should not be using
it.